requirements.txt
Flask==3.0.0
gunicorn==21.2.0
schedule==1.2.1
requests==2.31.0
python-dateutil==2.8.2
typing-extensions==4.9.0
python-dotenv==1.0.0 